requirements.txt
anyio==3.6.2
certifi==2022.12.7
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
mpmath==1.2.1
numpy==1.24.1
python-telegram-bot==20.0
rfc3986==1.5.0
sniffio==1.3.0
sympy==1.11.1
